soft/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
easyocr==1.7.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.2.0
psutil==5.9.8
pyclipper==1.3.0.post5
pyodbc==5.1.0
pyparsing==3.1.1
pyserial==3.5
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.13.2
shapely==2.0.3
six==1.16.0
sympy==1.12
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.0.114
urllib3==2.2.1